+++ readium/navigator/src/main/java/org/readium/r2/navigator/Navigator.kt
@@ -11,6 +11,7 @@ import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.StateFlow
 import org.readium.r2.navigator.media.MediaPlayback
+import org.readium.r2.shared.ExperimentalReadiumApi
 import org.readium.r2.shared.publication.Link
 import org.readium.r2.shared.publication.Locator
 import org.readium.r2.shared.publication.Publication
@@ -38,7 +39,6 @@ import kotlin.time.ExperimentalTime
  *   interactions such as tapping/clicking the edge of the page to skip to the next one should be
  *   implemented by the reading app, and not the navigator.
  */
-@OptIn(ExperimentalCoroutinesApi::class)
 interface Navigator {
 
     /**
@@ -104,10 +104,18 @@ interface NavigatorDelegate {
  * A navigator rendering the publication visually on-screen.
  */
 interface VisualNavigator : Navigator {
+
     /**
      * Current reading progression direction.
      */
     val readingProgression: ReadingProgression
+
+    /**
+     * Returns the [Locator] to the first content element that begins on the current screen.
+     */
+    @ExperimentalReadiumApi
+    suspend fun firstVisibleElementLocator(): Locator? =
+        currentLocator.value
 
     interface Listener : Navigator.Listener {
         /**
@@ -120,6 +128,33 @@ interface VisualNavigator : Navigator {
          * The [point] is relative to the navigator's view.
          */
         fun onTap(point: PointF): Boolean = false
+
+        /**
+         * Called when the user starts dragging the content, but nothing handled the event
+         * internally.
+         *
+         * The points are relative to the navigator's view.
+         */
+        @ExperimentalDragGesture
+        fun onDragStart(startPoint: PointF, offset: PointF): Boolean = false
+
+        /**
+         * Called when the user continues dragging the content, but nothing handled the event
+         * internally.
+         *
+         * The points are relative to the navigator's view.
+         */
+        @ExperimentalDragGesture
+        fun onDragMove(startPoint: PointF, offset: PointF): Boolean = false
+
+        /**
+         * Called when the user stops dragging the content, but nothing handled the event
+         * internally.
+         *
+         * The points are relative to the navigator's view.
+         */
+        @ExperimentalDragGesture
+        fun onDragEnd(startPoint: PointF, offset: PointF): Boolean = false
     }
 }